requirements.txt
python-telegram-bot==20.7
anthropic>=0.25.0
gspread>=6.0.0
google-auth>=2.28.0
python-dotenv>=1.0.0
